changes: new file structure function from funcs moved to api/nats and function from config moved to api/config
