+++ src/config.rs
@@ -3,9 +3,7 @@ use core::ffi::CStr;
 use pgrx::guc::*;
 use pgrx::prelude::*;
 
-use crate::connection::NATS_CONNECTION;
-use crate::funcs::do_panic_with_message;
-use crate::funcs::get_message;
+use crate::utils::format_message;
 
 // configs names
 const CONFIG_HOST: &str = "nats.host";
@@ -26,6 +24,7 @@ pub fn initialize_configuration() {
     GucContext::Userset,
     GucFlags::default(),
   );
+
   GucRegistry::define_int_guc(
     CONFIG_PORT,
     "port of rust service",
@@ -36,51 +35,10 @@ pub fn initialize_configuration() {
     GucContext::Userset,
     GucFlags::default(),
   );
+
   ereport!(
     PgLogLevel::INFO,
     PgSqlErrorCode::ERRCODE_SUCCESSFUL_COMPLETION,
-    get_message("initialize_configuration success!")
+    format_message("initialize_configuration success!")
   );
-}
-
-#[pg_extern]
-fn get_config(config_name: &str) -> Option<String> {
-  Spi::connect(|client| {
-    client
-      .select(
-        &format!("SELECT current_setting('{0}', true);", config_name),
-        None,
-        None,
-      )?
-      .first()
-      .get_one()
-  })
-  .ok()?
-}
-
-#[pg_extern]
-fn set_config(config_name: &str, config_value: &str) {
-  Spi::run(&format!("SET {} = {}", config_name, config_value)).unwrap_or_else(|_| {
-    do_panic_with_message(format!(
-      "Set configuration failed: <{}> -> <{}>",
-      config_name, config_value
-    ))
-  });
-
-  if config_name.to_lowercase().contains("nats.") {
-    NATS_CONNECTION.invalidate();
-  }
-}
-
-#[pg_extern]
-fn set_config_string(config_name: &str, config_value: &str) {
-  set_config(
-    config_name,
-    &format!("'{}'", config_value.replace("'", "''")),
-  );
-}
-
-#[pg_extern]
-fn reset_config(config_name: &str) {
-  set_config(config_name, "DEFAULT");
 }
