+++ src/lib.rs
@@ -3,9 +3,11 @@
 mod config;
 mod connection;
 mod errors;
-mod funcs;
 mod init;
 mod tests;
+mod utils;
+
+pub mod api;
 
 /// This module is required by `cargo pgrx test` invocations.
 /// It must be visible at the root of your extension crate.
