+++ src/tests.rs
@@ -3,11 +3,11 @@
 #[cfg(any(test, feature = "pg_test"))]
 #[pgrx::prelude::pg_schema]
 mod tests {
-  use crate::funcs;
+  use crate::api;
   use pgrx::prelude::*;
 
   #[pg_test]
   fn test_hello_pgnats() {
-    assert_eq!("Hello, pgnats!", funcs::hello_pgnats());
+    assert_eq!("Hello, pgnats!", api::hello_pgnats());
   }
 }
